requirements.txt
Flask==0.12.2
flask-cors==2.1.2
apache-beam[gcp]==2.6.0
gunicorn==19.7.1
google-cloud-datastore==1.6.0
google-cloud-dataflow==2.5.0
google-cloud-storage==1.8.0
google-cloud-bigquery[pandas]==1.3.0
requests==2.18.4
six==1.10.0